error: retire the module
`if_proto_bindings_missing` existed in this module all alone. The
function was previously used to determine if the required python
bindings for the jomiel protobuf messages were present.

Let it crash and burn; have Python interpreter print the traceback,
instead. Most users wouldn't see the human-friendly printed by the
function message anyway, because the release packager is responsible for
creating the bindings.