feat(log): log_init: add support for ident_must_exist (kwargs)
If ident_must_exist (str, the name of the logger identity specified in
the logger configuration YAML file) is defined, log_init() will make
sure that the identity exists among the loaded logger identities.
Otherwise, the process will exit with an error message.

Update also the example at the end of the file.